lista2

+++ automanager/src/main/java/com/autobots/automanager/controllers/AdressController.java
@@ -1,10 +1,13 @@
 package com.autobots.automanager.controllers;
 
 import com.autobots.automanager.entities.Adress;
-import com.autobots.automanager.models.AdressSelect;
+import com.autobots.automanager.models.AddLinkAdress;
 import com.autobots.automanager.models.AdressUpdate;
+import com.autobots.automanager.models.AdressSelect;
 import com.autobots.automanager.repositories.AdressRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,37 +19,74 @@ public class AdressController {
     private AdressRepository repository;
 
     @Autowired
-    private AdressSelect select;
+    private AdressSelect selector;
+
+    @Autowired
+    private AddLinkAdress addLink;
+
 
     @GetMapping("/enderecos")
-    public List<Adress> findAdress(){
-        List<Adress> adress = repository.findAll();
-        return adress;
+    public ResponseEntity<List<Adress>> findAdress(){
+        List<Adress> adresses = repository.findAll();
+        if (adresses.isEmpty()) {
+            ResponseEntity<List<Adress>> res = new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return res;
+        } else {
+            addLink.linkAdd(adresses);
+            ResponseEntity<List<Adress>> res = new ResponseEntity<>(adresses, HttpStatus.FOUND);
+            return res;
+        }
     }
 
     @GetMapping("/endereco/{id}")
-    public Adress findAdressById(@PathVariable Long id){
-        List<Adress> adress = repository.findAll();
-        return select.selector(adress, id);
+    public ResponseEntity<Adress> findAdressById(@PathVariable Long id){
+        List<Adress> adresses = repository.findAll();
+        Adress adress = selector.selector(adresses, id);
+        if (adress == null) {
+            ResponseEntity<Adress> res = new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return res;
+        } else {
+            addLink.linkAdd(adress);
+            ResponseEntity<Adress> res = new ResponseEntity<>(adress, HttpStatus.FOUND);
+            return res;
+        }
     }
 
     @PostMapping("/endereco")
-    public void createAdress(@RequestBody Adress adress) {
-        repository.save(adress);
+    public ResponseEntity<?> createAdress(@RequestBody Adress adress) {
+        HttpStatus status = HttpStatus.CONFLICT;
+        if (adress.getId() == null) {
+            repository.save(adress);
+            status = HttpStatus.CREATED;
+        }
+        return new ResponseEntity<>(status);
     }
 
 
     @PutMapping("/endereco")
-    public void updateAdress(@RequestBody Adress toUpdate) {
-        Adress endereco = repository.getById(toUpdate.getId());
-        AdressUpdate updateor = new AdressUpdate();
-        updateor.update(endereco, toUpdate);
-        repository.save(endereco);
+    public ResponseEntity<?> updateAdress(@RequestBody Adress updateInAdress) {
+        HttpStatus status = HttpStatus.CONFLICT;
+        Adress adress = repository.getById(updateInAdress.getId());
+        if (adress != null) {
+            AdressUpdate updator = new AdressUpdate();
+            updator.update(adress, updateInAdress);
+            repository.save(adress);
+            status = HttpStatus.OK;
+        } else {
+            status = HttpStatus.BAD_REQUEST;
+        }
+        return new ResponseEntity<>(status);
+
     }
 
     @DeleteMapping("/excluir")
-    public void deleteAdress(@RequestBody Adress exclude) {
+    public ResponseEntity<?> deleteAdress(@RequestBody Adress exclude) {
+        HttpStatus status = HttpStatus.BAD_REQUEST;
         Adress adress = repository.getById(exclude.getId());
-        repository.delete(adress);
+        if (adress != null) {
+            repository.delete(adress);
+            status = HttpStatus.OK;
+        }
+        return new ResponseEntity<>(status);
     }
 }
